requirements.txt
amqp==5.2.0
asgiref==3.7.2
aspose-words==23.12.0
async-timeout==4.0.3
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
celery-progress==0.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==4.2.8
django-celery-results==2.5.1
django-crispy-forms==2.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
drf-yasg2==1.19.4
fire==0.5.0
fonttools==4.47.0
htmldocx==0.0.6
humanize==4.9.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
inflection==0.5.1
itypes==1.2.0
flower==1.2.0
Jinja2==3.1.2
kombu==5.3.4
lxml==4.9.3
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.24.4
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pdf2docx==0.5.6
pdfkit==1.0.0
Pillow==10.1.0
proglog==0.1.10
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
sib-api-v3-sdk==7.6.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==2.4.0
tornado==6.4
tqdm==4.66.1
typing-extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
kafka-python==2.0.2
confluent-kafka==1.7.0
